requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8clear
djangorestframework>=3.0
pyjwt
python-dotenv